removal of convolver and switch to psf where required
