+++ autoarray/dataset/imaging/dataset.py
@@ -9,7 +9,6 @@ from autoarray.dataset.abstract.dataset import AbstractDataset
 from autoarray.dataset.grids import GridsDataset
 from autoarray.dataset.imaging.w_tilde import WTildeImaging
 from autoarray.structures.arrays.uniform_2d import Array2D
-from autoarray.operators.convolver import Convolver
 from autoarray.structures.arrays.kernel_2d import Kernel2D
 from autoarray.mask.mask_2d import Mask2D
 from autoarray import type as ty
@@ -30,7 +29,7 @@ class Imaging(AbstractDataset):
         noise_covariance_matrix: Optional[np.ndarray] = None,
         over_sample_size_lp: Union[int, Array2D] = 4,
         over_sample_size_pixelization: Union[int, Array2D] = 4,
-        pad_for_convolver: bool = False,
+        pad_for_psf: bool = False,
         use_normalized_psf: Optional[bool] = True,
         check_noise_map: bool = True,
     ):
@@ -77,7 +76,7 @@ class Imaging(AbstractDataset):
         over_sample_size_pixelization
             How over sampling is performed for the grid which is associated with a pixelization, which is therefore
             passed into the calculations performed in the `inversion` module.
-        pad_for_convolver
+        pad_for_psf
             The PSF convolution may extend beyond the edges of the image mask, which can lead to edge effects in the
             convolved image. If `True`, the image and noise-map are padded to ensure the PSF convolution does not
             extend beyond the edge of the image.
@@ -90,9 +89,9 @@ class Imaging(AbstractDataset):
 
         self.unmasked = None
 
-        self.pad_for_convolver = pad_for_convolver
+        self.pad_for_psf = pad_for_psf
 
-        if pad_for_convolver and psf is not None:
+        if pad_for_psf and psf is not None:
             try:
                 data.mask.derive_mask.blurring_from(
                     kernel_shape_native=psf.shape_native
@@ -175,25 +174,6 @@ class Imaging(AbstractDataset):
             over_sample_size_pixelization=self.over_sample_size_pixelization,
             psf=self.psf,
         )
-
-    @cached_property
-    def convolver(self):
-        """
-        Returns a `Convolver` from a mask and 2D PSF kernel.
-
-        The `Convolver` stores in memory the array indexing between the mask and PSF, enabling efficient 2D PSF
-        convolution of images and matrices used for linear algebra calculations (see `operators.convolver`).
-
-        This uses lazy allocation such that the calculation is only performed when the convolver is used, ensuring
-        efficient set up of the `Imaging` class.
-
-        Returns
-        -------
-        Convolver
-            The convolver given the masked imaging data's mask and PSF.
-        """
-
-        return Convolver(mask=self.mask, kernel=Kernel2D(values=self.psf._array, mask=self.psf.mask, header=self.psf.header))
 
     @cached_property
     def w_tilde(self):
@@ -370,7 +350,7 @@ class Imaging(AbstractDataset):
             noise_covariance_matrix=noise_covariance_matrix,
             over_sample_size_lp=over_sample_size_lp,
             over_sample_size_pixelization=over_sample_size_pixelization,
-            pad_for_convolver=True,
+            pad_for_psf=True,
         )
 
         dataset.unmasked = unmasked_dataset
@@ -463,7 +443,7 @@ class Imaging(AbstractDataset):
             noise_covariance_matrix=self.noise_covariance_matrix,
             over_sample_size_lp=self.over_sample_size_lp,
             over_sample_size_pixelization=self.over_sample_size_pixelization,
-            pad_for_convolver=False,
+            pad_for_psf=False,
             check_noise_map=False,
         )
 
@@ -511,7 +491,7 @@ class Imaging(AbstractDataset):
             over_sample_size_lp=over_sample_size_lp or self.over_sample_size_lp,
             over_sample_size_pixelization=over_sample_size_pixelization
             or self.over_sample_size_pixelization,
-            pad_for_convolver=False,
+            pad_for_psf=False,
             check_noise_map=False,
         )
