+++ autoarray/fixtures.py
@@ -110,12 +110,6 @@ def make_blurring_grid_2d_7x7():
     return aa.Grid2D.from_mask(mask=make_blurring_mask_2d_7x7())
 
 
-# CONVOLVERS #
-
-
-def make_convolver_7x7():
-    return aa.Convolver(mask=make_mask_2d_7x7(), kernel=make_psf_3x3())
-
 
 def make_image_7x7():
     return aa.Array2D.ones(shape_native=(7, 7), pixel_scales=(1.0, 1.0))
