+++ autoarray/operators/mock/mock_convolver.py
@@ -1,4 +1,4 @@
-class MockConvolver:
+class MockPSF:
     def __init__(self, operated_mapping_matrix=None):
         self.operated_mapping_matrix = operated_mapping_matrix
