+++ test_autoarray/inversion/inversion/imaging/test_inversion_imaging_util.py
@@ -181,7 +181,7 @@ def test__data_vector_via_w_tilde_data_two_methods_agree():
         shape_native=(7, 7), pixel_scales=mask.pixel_scales, sigma=1.0, normalize=True
     )
 
-    convolver = aa.Convolver(mask=mask, kernel=kernel)
+    psf = kernel
 
     pixelization = aa.mesh.Rectangular(shape=(20, 20))
 
@@ -203,7 +203,7 @@ def test__data_vector_via_w_tilde_data_two_methods_agree():
 
         mapping_matrix = mapper.mapping_matrix
 
-        blurred_mapping_matrix = convolver.convolve_mapping_matrix(
+        blurred_mapping_matrix = psf.convolve_mapping_matrix(
             mapping_matrix=mapping_matrix
         )
 
@@ -258,7 +258,7 @@ def test__curvature_matrix_via_w_tilde_two_methods_agree():
         shape_native=(7, 7), pixel_scales=mask.pixel_scales, sigma=1.0, normalize=True
     )
 
-    convolver = aa.Convolver(mask=mask, kernel=kernel)
+    psf = kernel
 
     pixelization = aa.mesh.Rectangular(shape=(20, 20))
 
@@ -282,7 +282,7 @@ def test__curvature_matrix_via_w_tilde_two_methods_agree():
         w_tilde=w_tilde, mapping_matrix=mapping_matrix
     )
 
-    blurred_mapping_matrix = convolver.convolve_mapping_matrix(
+    blurred_mapping_matrix = psf.convolve_mapping_matrix(
         mapping_matrix=mapping_matrix
     )
 
@@ -303,7 +303,7 @@ def test__curvature_matrix_via_w_tilde_preload_two_methods_agree():
         shape_native=(7, 7), pixel_scales=mask.pixel_scales, sigma=1.0, normalize=True
     )
 
-    convolver = aa.Convolver(mask=mask, kernel=kernel)
+    psf = kernel
 
     pixelization = aa.mesh.Rectangular(shape=(20, 20))
 
@@ -356,7 +356,7 @@ def test__curvature_matrix_via_w_tilde_preload_two_methods_agree():
             pix_pixels=pixelization.pixels,
         )
 
-        blurred_mapping_matrix = convolver.convolve_mapping_matrix(
+        blurred_mapping_matrix = psf.convolve_mapping_matrix(
             mapping_matrix=mapping_matrix
         )
